Добавлена проверка steam id из памяти игры по регулярному выражению, 17 чисел в строке. Исправлен баг чтения ника игрока из testStats.lua Оптимизирован код.

+++ APMMeasure.cpp
@@ -109,7 +109,7 @@ long APMMeasure::getAverageAPM() {
 	return computeAPM(getTotalActions(), GetTickCount()-absolute_starttick);
 }
 
-long APMMeasure::getTime()
+DWORD APMMeasure::getTime()
 {
     return GetTickCount()-absolute_starttick;
 }
